+++ duck/utils/safe_compare.py
@@ -11,6 +11,5 @@ def constant_time_compare(str_a: str, str_b: str):
     nomatter how short or long the 2 strings are, the time of comparing any kind of string is
     the same (constant), hence tackling timing attacks.
     """
-    assert isinstance(str_a, str) and isinstance(
-        str_b, str), "Only strings are allowed for both the arguments"
+    assert isinstance(str_a, str) and isinstance(str_b, str), "Only strings are allowed for both the arguments"
     return hmac.compare_digest(str_a, str_b)